backend/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1        # no [standard]
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.0
